% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dft_acf}
\alias{dft_acf}
\title{Discrete Fourier Transformation for Autocovariance Function}
\usage{
dft_acf(x)
}
\arguments{
\item{x}{A \code{cx_vec}.}
}
\value{
A \code{vec} containing the ACF.
}
\description{
Calculates the autovariance function (ACF) using Discrete Fourier Transformation.
}
\details{
This implementation is 2x as slow as Rs. 
Two issues: 1. memory resize and 2. unoptimized fft algorithm in arma.
Consider piping back into R and rewrapping the object. (Decrease of about 10 microseconds.)
}
\examples{
x=rnorm(100)
dft_acf(x)
}

