% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diff_cpp}
\alias{diff_cpp}
\title{Lagged Differences in Armadillo}
\usage{
diff_cpp(x, lag, differences)
}
\arguments{
\item{x}{A \code{vec} that is the time series}

\item{lag}{A \code{unsigned int} that indicates the lag}

\item{differences}{A \code{dif} that indicates how many differences should be taken}
}
\value{
A \code{vector} containing the differenced time series.
}
\description{
Returns the ith difference of a time series of rth lag.
}
\examples{
x = rnorm(10000, 0, 1)
diff_cpp(x,1,1)
}
\author{
JJB
}

