% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{field_to_matrix}
\alias{field_to_matrix}
\title{Transform an Armadillo field<vec> to a matrix}
\usage{
field_to_matrix(x)
}
\arguments{
\item{x}{A \code{field<vec>}.}
}
\value{
A \code{mat} containing the field elements within a column.
}
\description{
Unlists vectors in a field and places them into a matrix
}
\examples{
x=rnorm(100)
field_to_matrix(modwt_cpp(x))
}
\author{
JJB
}

