% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_elements}
\alias{get_elements}
\title{Subset Non-connected Regions}
\usage{
get_elements(x, row_ind, col_ind)
}
\arguments{
\item{x}{A \code{matrix} of dimensions M x N}

\item{row_ind}{A \code{unsigned int vec} that contains the row indices within \eqn{[0,M-1]}.}

\item{col_ind}{A \code{unsigned int vec} that contains the column indices within \eqn{[0,N-1]}.}
}
\value{
A \code{vec} with each element listed according to index specification.
}
\description{
Replicates the subset functionality of a matrix in R.
}
\examples{
# Generate a Matrix
m = matrix(1:12, nrow = 4)

# Select Non-connect regions
row_index = c(1, 2, 1)
col_index = c(2, 2, 3)

# Subset in R
m[cbind(row_index, col_index)]

# Subset with Armadillo
get_elements(m, row_index - 1, col_index - 1)
}
\author{
JJB
}

