% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2arma_package.R
\docType{package}
\name{r2arma-package}
\alias{r2arma}
\alias{r2arma-package}
\title{R-based using the Armadillo C++ Matrix Library}
\description{
Provides C++ headers that contain implementations of some base R
functions.
}

