% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rev_col_subset}
\alias{rev_col_subset}
\title{Reverse Subset Column}
\usage{
rev_col_subset(x, start, end)
}
\arguments{
\item{x}{A \code{matrix} of dimensions M x N}

\item{start}{A \code{unsigned int} that indicates the starting column.}

\item{end}{A \code{unsigned int} that indicates the ending column.}
}
\value{
A \code{matrix} with matrix rows displayed in reverse order
}
\description{
Subsets the column by going from high indices to low (the reverse of the supported practice)
}
\details{
Consider a vector x=[[1,2],[3,4]].
By setting \code{start=1} and \code{end=0}, the function would output x=[[2,1],[4,1]].
Start and end must be valid C++ matrix locations. (e.g. matrix cols start at 0 and not 1)
}
\examples{
x = matrix(c(1,2,3,4), nrow = 2,byrow = TRUE)
rev_col_subset(x, 1, 0)
}
\author{
JJB
}
