% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reverse_vec}
\alias{reverse_vec}
\title{Reverse Armadillo Vector}
\usage{
reverse_vec(x)
}
\arguments{
\item{x}{A \code{column vector} of length N}
}
\value{
x A \code{column vector} with its contents reversed.
}
\description{
Reverses the order of an Armadillo Vector
}
\details{
Consider a vector x=[1,2,3,4,5], the function would output x=[5,4,3,2,1].
}
\examples{
x = 1:5
reverse_vec(x)
}
\author{
JJB
}

