% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/seq.cpp, include/seq.h
\name{seq_along_cpp}
\alias{seq_along_cpp}
\title{Generate a sequence of values ranging from vector start to vector end}
\usage{
seq_along_cpp(along_with)
}
\arguments{
\item{along_with}{A \code{vec} with length \eqn{n}.}
}
\value{
A \code{vector} of length \eqn{n} that contains integer values in \eqn{[0, n - 1]}.
}
\description{
Creates a vector containing a sequence of values starting at the beginning and going to the end.
}
\examples{
# Call with the following data:
seq_along_cpp(1:10)
seq_along_cpp(5:10)
}
\seealso{
\code{\link{seq}}
}
\author{
James J Balamuta
}
