% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/seq.cpp, include/seq.h
\name{seq_default}
\alias{seq_default}
\title{Generate a sequence of values ranging from a to b given a fixed amount of points.}
\usage{
seq_default(from, to, length_out)
}
\arguments{
\item{from}{A \code{long double} that denotes the starting point.}

\item{to}{A \code{long double} indicating the end point.}

\item{length_out}{A \code{long unsigned int} that denotes the number of points to use.}
}
\value{
A \code{vector} containing values moving from \eqn{[from,to]}.
}
\description{
Creates a vector containing a sequence of values starting at the initial point and going to the terminal point.
}
\examples{
# Call with the following data:
seq_default(1, 2, 10)
seq_default(2, 1, 10)
}
\seealso{
\code{\link{seq}}
}
\author{
Anthony R. Colombo, James J Balamuta
}
