% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq_default_a}
\alias{seq_default_a}
\title{Generate a sequence of values ranging from -a to a given a fixed amount of points.}
\usage{
seq_default_a(a, length_out)
}
\arguments{
\item{a}{A \code{long double} that denotes the starting point.}

\item{length_out}{A \code{long unsigned int} that denotes the number of points to use.}
}
\value{
A \code{vector} of length \code{length_out} that contains values in \eqn{[-a,a]}.
}
\description{
Creates a vector containing a sequence of values starting at the initial point and going to the terminal point.
}
\examples{
# Call with the following data:
seq_default_a(1, 10)
seq_default_a(1, 10)
}
\author{
Anthony R. Colombo, James J Balamuta
}
\seealso{
\code{\link{seq}}
}

