% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/seq.cpp, include/seq.h
\name{seq_len_cpp}
\alias{seq_len_cpp}
\title{Generate a sequence of values ranging from 0 to n-1}
\usage{
seq_len_cpp(length_out)
}
\arguments{
\item{length_out}{An \code{long unsigned int} that denotes the number of points to use.}
}
\value{
A \code{vector} of length \code{length_out} that contains integer values in \eqn{[0, length_out - 1]}.
}
\description{
Creates a vector containing a sequence of values starting at the initial point and going to the terminal point.
}
\examples{
# Call with the following data:
seq_len_cpp(2)
seq_len_cpp(4)
}
\author{
James J Balamuta
}
\seealso{
\code{\link{seq}}
}

