% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_field_vec}
\alias{sum_field_vec}
\title{Accumulation of Armadillo field<vec>}
\usage{
sum_field_vec(x)
}
\arguments{
\item{x}{A \code{field<vec>}.}
}
\value{
An \code{mat} containing the field elements within a column.
}
\description{
Sums vectors in a field into a single variable.
}
\examples{
x=rnorm(100)
field_to_matrix(modwt_cpp(x))
}
\author{
JJB
}

